Move text utility function to a new file
There are enough functions now that they should be split from the
poly_line_text_reducer file.
